make type of matrix String
this allows the user to type any data that can be converted to text, as long as there are no spaces in the value of a cell.
since numbers are also stringifyable and we don't do any calculations with the matrix elements, they will be displayed normally on screen, so the usual case isn't affected by the change.